Bugfix for SqlConnectionPool + Moved to Cave.Data.Sql + Fixed usage of databaseName for global context
